pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "eda_openenv"
version = "1.0.0"
description = "A real-world RL environment for EDA pipeline agents, built on the OpenEnv standard."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Rahmath", email = "rahmathits@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    # openenv-core — auto-generates HTTP endpoints via create_app()
    "openenv-core",
    # API server
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    # Data science
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    # Inference script
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",   # for FastAPI TestClient
]

[project.urls]
Homepage   = "https://huggingface.co/spaces/Rahmath1/OpenEnv"
Repository = "https://github.com/rahmathits/OpenEn-Hackathon"

[project.scripts]
# Run server directly: uv run server
server = "server.app:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["server*", "tools*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.yaml", "*.txt"]

# ─────────────────────────────────────────
# uv configuration
# ─────────────────────────────────────────
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]